pswf_1

+++ firmware/app/mrd/voiceserviceinterface.h
@@ -35,7 +35,7 @@ public:
     void tuneSquelch(uint8_t value);
     void TuneAudioLevel(uint8_t volume_level);
     void TurnAGCMode(uint8_t mode);
-    void TurnFHSSMode(uint8_t mode, float LCODE, float RN_KEY,float COM_N, float FREQ);
+    void TurnFHSSMode(uint8_t mode, int LCODE, int RN_KEY,int COM_N, uint32_t FREQ);
